Non-portable ncurses term init fixes for the Apple Silicon.
Apparently variadic substitutions are somewhat broken on macos@arm64
so strict, albeit non-portable functions declarations resolve the issue.
Solution is working for the linux/mac arm64/amd64 but probably not for
the Solaris and some other; therefore it is considered a hack till some
better idea appears.
One of the ideas (props to @Nei@Freenode) is to search for ncursesw/term.h
header to include from the buildsystem; that way it works and portable but
on m1 builds only with homebrew ncurses package, not MacOS SDK one.